+++ src/aspire/source/micrograph.py
@@ -155,7 +155,7 @@ class ArrayMicrographSource(MicrographSource):
         )
 
         # We're already backed by an array, access it directly.
-        self._data = micrographs.astype(self.dtype, copy=False)
+        self._data = micrographs.astype(self.dtype, copy=None)
 
     def _images(self, indices):
         """
